Fix for 5-argstostr-- changed vars to unsigned int

+++ 0x0B-malloc_free/5-argstostr.c
@@ -11,17 +11,17 @@
 
 char *argstostr(int ac, char **av)
 {
-	int a, i, c = 0, s = 0;
+	unsigned int a, i, c = 0, s = 0;
 	char *str;
 
 	if (ac == 0 || av == NULL)
 		return (NULL);
-	for (a = 0; a < ac; a++)
+	for (a = 0; a < (unsigned int)ac; a++)
 		s += sizeof(av[a] + 1);
 	str = malloc(s + 1);
 	if (str == NULL)
 		return (NULL);
-	for (a = 0; a < ac; a++)
+	for (a = 0; a < (unsigned int)ac; a++)
 	{
 		i = 0;
 		while (av[a][i])
